BEE STUFF! (And Nether bricks, but that's not why you're here)
Don't have the entities yet, but I did everything else. The honeycomb block still needs some work because it it folded weird. I also noticed that the Red Nether Brick texture was trash and fixed it. Also, the Nether brick textures were missing from my files, but was in the .minecraft file for some reason.